+++ test/e2e/storage/storage.cases.js
@@ -1,7 +1,9 @@
 StorageSelectorScenarios = require('./cases/storage-selector.js');
 NewFolderModalScenarios = require('./cases/new-folder.js');
 UploadScenarios = require('./cases/upload.js');
+HomeScenarios = require('./cases/home.js');
 
 storageSelectorScenarios = new StorageSelectorScenarios();
 newFolderModalScenarios = new NewFolderModalScenarios();
 uploadScenarios = new UploadScenarios();
+homeScenarios = new HomeScenarios();
