+++ web/index.html
@@ -206,8 +206,10 @@
   <script src="scripts/storage/directives/dtv-storage-selector.js"></script>
   <script src="scripts/storage/directives/dtv-storage-file-select.js"></script>
   
+  <script src="scripts/storage/controllers/ctr-files-list.js"></script>
   <script src="scripts/storage/controllers/ctr-storage-selector-modal.js"></script>
   <script src="scripts/storage/controllers/ctr-new-folder-modal.js"></script>
+  <script src="scripts/storage/controllers/ctr-storage.js"></script>
   <script src="scripts/storage/controllers/ctr-subscription-status.js"></script>
   <script src="scripts/storage/controllers/ctr-upload.js"></script>
