+++ web/partials/storage/files-list.html
@@ -1,0 +1,94 @@
+<div class="content-box half-top" ng-show="!trialAvailable">
+  <div class="scrollable-list" 
+  rv-spinner rv-spinner-key="storage-selector-loader"
+  rv-spinner-start-active="1" >
+  <table id="storageFileList" class="table-2 table-hover table-selector" ng-class="storageFactory.storageFull || storageFactory.isMultipleFileSelector() ? 'multiple-selector' : 'single-selector'" ng-show="statusDetails.code!==202 && statusDetails.code!==404 && isFileListVisible()">
+    <thead>
+      <tr>
+        <th class="col-sm-6">
+          <input type="checkbox" class="add-right" ng-model="selectAll">
+          <label ng-click="fileSelectorFactory.selectAllCheckboxes(query)" ng-show="storageFactory.storageFull || storageFactory.isMultipleFileSelector()"></label>
+          <a href="" ng-click="orderByAttribute = fileNameOrderFunction; reverseSort = !reverseSort">
+            <span translate="common.file-name" ></span>
+            <span ng-show="orderByAttribute==fileNameOrderFunction">
+              <span ng-show="reverseSort">
+                <i class="fa fa-long-arrow-up"></i>
+              </span>
+              <span ng-show="!reverseSort">
+                <i class="fa fa-long-arrow-down"></i>
+              </span>
+            </span>
+          </a>
+        </th>
+        <th class="col-sm-2 hidden-xs" ng-click="orderByAttribute = fileExtOrderFunction; reverseSort = !reverseSort">
+          <a href=""><span translate="common.file-type"></span>
+            <span ng-show="orderByAttribute==fileExtOrderFunction">
+              <span ng-show="reverseSort">
+                <i class="fa fa-long-arrow-up"></i>
+              </span>
+              <span ng-show="!reverseSort">
+                <i class="fa fa-long-arrow-down"></i>
+              </span>
+            </span>
+          </a>
+        </th>
+        <th class="col-sm-2 visible-lg" ng-click="orderByAttribute = fileSizeOrderFunction; reverseSort = !reverseSort">
+          <a href=""><span translate="common.file-size"></span>
+            <span ng-show="orderByAttribute==fileSizeOrderFunction">
+              <span ng-show="reverseSort">
+                <i class="fa fa-long-arrow-up"></i>
+              </span>
+              <span ng-show="!reverseSort">
+                <i class="fa fa-long-arrow-down"></i>
+              </span>
+            </span>
+          </a>
+        </th>
+        <th class="col-sm-2 visible-lg" ng-click="orderByAttribute = dateModifiedOrderFunction; reverseSort = !reverseSort">
+          <a href=""><span translate="common.file-date-mod"></span>
+            <span ng-show="orderByAttribute==dateModifiedOrderFunction">
+              <span ng-show="reverseSort">
+                <i class="fa fa-long-arrow-up"></i>
+              </span>
+              <span ng-show="!reverseSort">
+                <i class="fa fa-long-arrow-down"></i>
+              </span>
+            </span>
+          </a>
+        </th>
+      </tr>
+    </thead>
+    <tbody>
+      <tr class="clickable-row"
+      ng-click="fileClick(file);"
+      ng-class="{'active': file.isChecked, 'blocked-file': file.isThrottled, 'back-btn': file.currentFolder, 'no-select-row': file.currentFolder || storageFactory.fileIsTrash(file) || (storageFactory.fileIsFolder(file) && !storageFactory.storageFull) }"
+      ng-repeat="file in filesDetails.files | filter:search.query | orderBy:orderByAttribute:reverseSort track by $index">
+      <td ng-if="storageFactory.fileIsFolder(file) && (!storageFactory.fileIsTrash(file) || storageFactory.storageFull)">
+        <span class="folder">{{file.name | fileNameFilter:storageFactory.folderPath}}</span>
+      </td>
+      <td ng-if="!storageFactory.fileIsFolder(file)" class="col-sm-6">
+        <span class="file" ng-class='{"throttled-type": file.isThrottled && !storageFactory.fileIsFolder(file)}'>
+          <a style="white-space: nowrap;overflow: hidden;text-overflow: ellipsis;" href="" data-toggle="tooltip" title="{{file.name | fileNameFilter:currentDecodedFolder}}">
+            <i class="fa" ng-class='{"fa-exclamation": file.isThrottled && !storageFactory.fileIsFolder(file), "icon-red": file.isThrottled && !storageFactory.fileIsFolder(file)}'></i>
+            {{file.name | fileNameFilter:storageFactory.folderPath}}
+          </a>
+        </span>
+        <div ng-show="file.showThrottledCallout && !storageFactory.fileIsFolder(file)" class="callout callout-red" ng-click="file.showThrottledCallout=false">
+          <span class="arrow"></span>
+          <strong>File Is Blocked</strong>
+          <p>This file is being requested too many times and is blocked. If you have a <a href="http://support.risevision.com" target="_blank">support subscription</a> please open a ticket for assistance with resolving this.  Otherwise, please post in the <a href="http://community.risevision.com" target="_blank">community</a> for help.</p>
+        </div>
+      </td>
+      <td ng-if="!fileIsFolder(file)" class="col-sm-2 hidden-xs">
+        <span> {{file.name | fileTypeFilter}}</span>
+      </td>
+      <td ng-if="!fileIsFolder(file)" class="col-sm-2 visible-lg">
+        <span> {{file.size | fileSizeFilter}}</span>
+      </td>
+      <td ng-if="!fileIsFolder(file)" class="col-sm-2 visible-lg">
+        <span> {{file.updated.value | date:'d-MMM-yyyy h:mm a'}}</span>
+      </td>
+    </tr>
+  </tbody>
+</table>
+</div>
