+++ web/scripts/storage/controllers/ctr-storage.js
@@ -1,0 +1,7 @@
+'use strict';
+angular.module('risevision.storage.controllers')
+  .controller('StorageController', ['$scope', 'storageFactory',
+    function ($scope, storageFactory) {
+      storageFactory.storageFull = true;
+    }
+  ]);
